requirements.txt
# ============================================
# DEPENDÊNCIAS DO SISTEMA AUTOPEÇAS
# ============================================
# Versão: 1.0.1
# Python: 3.8+
# Atualizado em: 24/11/2025

# ============================================
# DJANGO E CORE
# ============================================
Django>=5.2.0,<5.3.0
django-environ==0.11.2        # Gerenciamento de variáveis de ambiente
python-decouple==3.8          # Configurações do projeto

# ============================================
# API REST
# ============================================
djangorestframework==3.14.0   # Framework para API REST
django-filter==23.5           # Filtros para DRF
django-cors-headers==4.3.1    # CORS para API
drf-yasg==1.21.7             # Documentação automática da API (Swagger)

# ============================================
# BANCO DE DADOS
# ============================================
psycopg2-binary==2.9.9       # PostgreSQL adapter (produção)
# mysqlclient==2.2.1         # MySQL adapter (se necessário)

# ============================================
# IMAGENS E ARQUIVOS
# ============================================
Pillow==10.1.0               # Processamento de imagens
python-magic==0.4.27         # Detecção de tipo de arquivo

# ============================================
# EXCEL E DOCUMENTOS
# ============================================
openpyxl==3.1.2              # Leitura/escrita de arquivos Excel
xlrd==2.0.1                  # Leitura de arquivos .xls antigos
xlwt==1.3.0                  # Escrita de arquivos .xls
pandas==2.1.4                # Análise e manipulação de dados
reportlab==4.0.7             # Geração de PDFs

# ============================================
# FORMULÁRIOS E VALIDAÇÃO
# ============================================
django-crispy-forms==2.1      # Formulários Bootstrap
crispy-bootstrap5==2024.2     # Templates Bootstrap 5
python-cpf==1.5.0            # Validação de CPF
python-cnpj==2.1.0           # Validação de CNPJ
validate-docbr==1.10.0       # Validação de documentos brasileiros

# ============================================
# AUTENTICAÇÃO E SEGURANÇA
# ============================================
django-allauth==0.57.0       # Autenticação completa
djangorestframework-simplejwt==5.3.1  # JWT para API
argon2-cffi==23.1.0          # Hashing de senhas seguro
cryptography==41.0.7         # Criptografia

# ============================================
# CACHE E PERFORMANCE
# ============================================
redis==5.0.1                 # Cliente Redis
django-redis==5.4.0          # Cache backend Redis
hiredis==2.3.2               # Parser Redis C (performance)

# ============================================
# TAREFAS ASSÍNCRONAS
# ============================================
celery==5.3.4                # Tarefas assíncronas
django-celery-beat==2.5.0    # Agendamento de tarefas
django-celery-results==2.5.1 # Armazenamento de resultados

# ============================================
# MONITORAMENTO E LOGS
# ============================================
django-debug-toolbar==4.2.0  # Debug toolbar (apenas dev)
sentry-sdk==1.39.1           # Monitoramento de erros
python-json-logger==2.0.7    # Logs estruturados em JSON

# ============================================
# TESTES
# ============================================
pytest==7.4.3                # Framework de testes
pytest-django==4.7.0         # Integração pytest-django
pytest-cov==4.1.0            # Cobertura de testes
factory-boy==3.3.0           # Factories para testes
faker==20.1.0                # Dados fake para testes
coverage==7.3.3              # Análise de cobertura

# ============================================
# QUALIDADE DE CÓDIGO
# ============================================
flake8==6.1.0                # Linter Python
black==23.12.1               # Formatador de código
isort==5.13.2                # Organizador de imports
pylint==3.0.3                # Análise estática de código
mypy==1.7.1                  # Type checker

# ============================================
# UTILIDADES
# ============================================
python-dateutil==2.8.2       # Manipulação de datas
pytz==2023.3.post1           # Fusos horários
requests==2.31.0             # HTTP requests
urllib3==2.1.0               # HTTP client
certifi==2023.11.17          # Certificados SSL

# ============================================
# INTEGRAÇÕES
# ============================================
# python-correios==8.3.0     # Integração Correios (se necessário)
# pycep-correios==5.0.0      # Busca CEP
# pyboleto==0.3.5            # Geração de boletos
# pynfe==0.1.0               # Nota Fiscal Eletrônica

# ============================================
# SERVIDOR DE PRODUÇÃO
# ============================================
gunicorn==21.2.0             # WSGI server
whitenoise==6.6.0            # Servir arquivos estáticos
dj-database-url==2.1.0       # Parse database URLs

# ============================================
# UTILITÁRIOS DE DESENVOLVIMENTO
# ============================================
ipython==8.18.1              # Shell interativo melhorado
django-extensions==3.2.3     # Comandos úteis para Django
werkzeug==3.0.1              # Utilitários WSGI

# ============================================
# DOCUMENTAÇÃO
# ============================================
Sphinx==7.2.6                # Gerador de documentação
sphinx-rtd-theme==2.0.0      # Tema Read the Docs

# ============================================
# NOTAS DE INSTALAÇÃO
# ============================================
# 
# Para instalar todas as dependências:
#   pip install -r requirements.txt
#
# Para instalar apenas produção:
#   pip install -r requirements.txt --no-dev
#
# Para instalar em sistemas sem privilégios (Linux):
#   pip install -r requirements.txt --break-system-packages
#
# Para atualizar todas as dependências:
#   pip install -r requirements.txt --upgrade
#
# Para gerar requirements.txt atual:
#   pip freeze > requirements.txt
#
# ============================================
# COMPATIBILIDADE
# ============================================
# Python: 3.8, 3.9, 3.10, 3.11, 3.12
# Django: 5.2.x
# PostgreSQL: 12+
# Redis: 6+
# Sistema Operacional: Windows, Linux, macOS
#
# ============================================
# AMBIENTES
# ============================================
#
# DESENVOLVIMENTO:
# - Usar SQLite
# - DEBUG = True
# - django-debug-toolbar habilitado
#
# PRODUÇÃO:
# - Usar PostgreSQL
# - DEBUG = False
# - Configurar ALLOWED_HOSTS
# - Usar gunicorn + nginx
# - Configurar cache Redis
# - Configurar Celery para tarefas
# - Configurar Sentry para logs
#
# ============================================
